Web/Controllers/AuthController.cs
using Application.Users.Login;
using Application.Users.Register;
using MediatR;
using Microsoft.AspNetCore.Mvc;
using Web.Requests;

namespace Web.Controllers;

[Route("api/auth")]
public class AuthController: ControllerBase
{
    private readonly ISender _sender;

    public AuthController(ISender sender)
    {
        _sender = sender;
    }

    [HttpPost("login")]
    public async Task<IActionResult> LoginAsync([FromBody] LoginRequest request, CancellationToken cancellationToken)
    {
        var command = new LoginCommand(request.Email, request.Password);

        var result = await _sender.Send(command, cancellationToken);

        return result.IsSuccess ? Ok(result.Value) : BadRequest(result.Error.Message);
    }
    
    [HttpPost("register")]
    public async Task<IActionResult> RegisterAsync([FromBody] RegisterRequest request, CancellationToken cancellationToken)
    {
        var command = new RegisterCommand(request.Email, request.Password);

        var result = await _sender.Send(command, cancellationToken);

        return result.IsSuccess ? Created() : BadRequest(result.ValidationErrors);
    }
}